refactor: merge Puppeteer and PuppeteerNode
