+++ docs/api/puppeteer.puppeteer.trimcache.md
@@ -1,15 +1,15 @@
 ---
-sidebar_label: PuppeteerNode.trimCache
+sidebar_label: Puppeteer.trimCache
 ---
 
-# PuppeteerNode.trimCache() method
+# Puppeteer.trimCache() method
 
 Removes all non-current Firefox and Chrome binaries in the cache directory identified by the provided Puppeteer configuration. The current browser version is determined by resolving PUPPETEER_REVISIONS from Puppeteer unless `configuration.browserRevision` is provided.
 
 #### Signature:
 
 ```typescript
-class PuppeteerNode {
+class Puppeteer {
   trimCache(): Promise<void>;
 }
 ```
