+++ test/src/cdp/extensions.spec.ts
@@ -7,7 +7,7 @@
 import path from 'path';
 
 import expect from 'expect';
-import type {PuppeteerLaunchOptions} from 'puppeteer-core/internal/node/PuppeteerNode.js';
+import type {PuppeteerLaunchOptions} from 'puppeteer-core/internal/node/Puppeteer.js';
 
 import {getTestState, launch} from '../mocha-utils.js';
